feat(function): RenderLoom split → RenderBroken + RenderFixed (one mode each)
Decompose §5's bundled RenderLoom into two single-mode widgets. RenderBroken
shows a callback tethered to render 0 — every tick reads the stale 0, screen
sticks at 1. RenderFixed shows the updater form (c => c + 1) with no tether,
the screen counts up cleanly. The broken-vs-fixed contrast moves into the
connecting prose between them. ModeToggle deleted (last caller). buildTimeline
hoisted to module scope as TICKS const in each new widget — no more useMemo
cargo-cult, no allocation per mount. AnimatePresence dropped (its only purpose
was branch-swap exits that no longer happen); motion.g mount-in fade preserved
on the tether and binding-cell groups. memo wrap on the narrow components
preserved. Container-query CSS classes renamed in lockstep.
