+++ app/posts/the-function-that-remembered/widgets/index.ts
@@ -1,5 +1,6 @@
 export { LoopLet } from "./LoopLet";
 export { LoopVar } from "./LoopVar";
 export { PredictReveal } from "./PredictReveal";
-export { RenderLoom } from "./RenderLoom";
+export { RenderBroken } from "./RenderBroken";
+export { RenderFixed } from "./RenderFixed";
 export { TetherScope } from "./TetherScope";
